move reference_links() to postutils
create testutils by introducing DatabaseTest
class, to be inherited by any Test* class which
requires the use of a database for testing
purposes.

Make reference_links() accept the Post model as
its first argument. The reason reference_links()
was moved to postutils was for clearer separation
of module responsibilities. It didn't seem right
that so much post transmorgification ocurred in
the models module/in Post model.

Remove `test_models` because the only thing it
was testing was reference_links() which is now
a postutils function.